+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## master
+
+* Adds source map upload script support for environment variables use inside Info.plist
 
 ## v9.1.9 (2020-10-01)